+++ govlearn-api/src/main/java/com/unimuenster/govlearnapi/group/service/dto/GroupDTO.java
@@ -1,7 +1,5 @@
 package com.unimuenster.govlearnapi.group.service.dto;
 
-import com.unimuenster.govlearnapi.course.service.dto.CourseDTO;
-import com.unimuenster.govlearnapi.user.service.dto.UserDTO;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -11,8 +9,8 @@ import java.util.List;
 @Setter
 public class GroupDTO {
 
-    List<Long> members;
-    List<Long> courses;
-    Long creatorId;
+    List<Integer> members;
+    List<Integer> courses;
+    Integer creatorId;
     String name;
 }
